+++ mazesolver/strings.py
@@ -4,7 +4,9 @@ This file contains all long strings used by the program
 help_message = """\
 Valid Commands
 ---------------
--h or --help    -  Prints out this help page
+-h or --help    -  Prints this help page and exits
+-v or --version -  Prints the version of the program and exits
+
 -i or --input   -  Input maze file
 -o or --output  -  Output directory
 
@@ -19,4 +21,6 @@ Github: https://github.com/exciteabletom
 *Licensed under GPLv3* [https://www.gnu.org/licenses/gpl-3.0.en.html]
 """
 
-version="1.2.2-1"
+version="1.2.2.post1"
+version_long=f"mazesolver {version}"
+
